+++ sources/webapp/scripts/ui/speech/synthesis.js
@@ -27,6 +27,17 @@
 /* global SpeechSynthesisUtterance */
 
 /**
+  * A promise fulfilled when a text is spoken.
+  *
+  * @promise SpeakPromise
+  * @fulfill {undefined} Fulfill empty when a text is spoken.
+  * @reject {external:String} A {@link https://developer.mozilla.org/en-US/docs/Web/API/SpeechSynthesisErrorEvent/error|error code} from the Web Speech API.
+  * @reject {external:Error} A Error if no offline avaliable voice is installed.
+  */
+
+const LOGGER = window.podcatcher.configuration.logging.createLogger('h5p/speech/synthesis')
+
+/**
   * Returns true if the {@link https://developer.mozilla.org/en-US/docs/Web/API/Web_Speech_API|Web Speech API} ist supported by this platform.
   * @static
   * @returns {boolean} True if the Web Speech API is supported by this platform.
@@ -38,7 +49,7 @@ export function isSupported () {
 /**
   * Returns a list of voices for the Web Speech API that are offline available.
   * @static
-  * @returns {SpeechSynthesisVoice[]} Voices that are offline available.
+  * @returns {external:SpeechSynthesisVoice[]} Voices that are offline available.
   */
 export function getVoices () {
   if (isSupported) {
@@ -50,7 +61,7 @@ export function getVoices () {
         offlineVoices.push(voice)
       }
     }
-    return offlineVoices
+    return offlineVoices.sort(voiceComparator)
   } else {
     throw new Error('Speech Synthesis isn\'t supported by this platform.')
   }
@@ -61,16 +72,18 @@ export function getVoices () {
   * @param {external:String[]} [favoriteVoiceNames=[]] Name of voices to favorize speaking a text.
   * @param {number} [rate=1] Rate to speak a text width.
   * @param {number} [pitch=1] Pitch to speak a text width.
+  * @param {number} [volume=1] The volume to speak. A number between 0 and 1.
   */
 export class Synthesiser {
   /**
     * @constructs
     */
-  constructor (favoriteVoiceNames, rate, pitch) {
+  constructor (favoriteVoiceNames, rate, pitch, volume) {
     if (isSupported) {
       this.favoriteVoices = favoriteVoiceNames || []
       this.rate = rate || 1
       this.pitch = pitch || 1
+      this.volume = volume || 1
     } else {
       throw new Error('Speech synthesis isn\'t supported by this platform.')
     }
@@ -94,6 +107,15 @@ export class Synthesiser {
     return this._rate
   }
 
+  set volume (value) {
+    // TODO check number
+    // TODO check range 0 to 1
+    this._volume = value
+  }
+  get volume () {
+    return this._volume
+  }
+
   set favoriteVoices (favoriteVoiceNames) {
     // TODO check Array
     // TODO check String-Items
@@ -106,26 +128,64 @@ export class Synthesiser {
   /** Speak a text in the given language.
     * @param {external:String} text Text to speak
     * @param {external:String} [lang=en] Language to speak in.
-    * @returns {undefined}
+    * @returns {SpeakPromise} A Promise that fulfill when the given text is spoken.
     */
   speak (text, lang) {
-    const synthesiser = window.speechSynthesis
-    const utterance = new SpeechSynthesisUtterance(text)
-
-    lang = lang || 'en'
-
     // Select a voice for the given language
+    let selectedVoice = null
     for (let i = 0; i < getVoices().length; i++) {
       const voice = getVoices()[i]
       if (voice.lang.indexOf(lang) === 0) {
-        utterance.voice = voice
+        selectedVoice = voice
         if (this.favoriteVoices.includes(voice.name)) {
           break
         }
       }
     }
-    utterance.pitch = this.pitch
-    utterance.rate = this.rate
-    synthesiser.speak(utterance)
+
+    if (selectedVoice) {
+      const synthesiser = window.speechSynthesis
+      const utterance = new SpeechSynthesisUtterance(text)
+
+      lang = lang || 'en'
+
+      utterance.voice = selectedVoice
+      utterance.pitch = this.pitch
+      utterance.rate = this.rate
+      utterance.volume = this.volume
+      return new Promise((resolve, reject) => {
+        utterance.addEventListener('error', (/* SpeechSynthesisErrorEvent */event) => {
+          LOGGER.error(`An error has occurred with the speech synthesis: ${event.error}`)
+          reject(event.error)
+        })
+        utterance.addEventListener('end', (/* SpeechSynthesisEvent */event) => {
+          LOGGER.debug(`Utterance has finished being spoken after ${event.elapsedTime} milliseconds.`)
+          resolve()
+        })
+        synthesiser.speak(utterance)
+        LOGGER.debug(`Speech text with voice ${utterance.voice.name}, pith ${utterance.pitch}, volume ${Math.floor(utterance.volume * 100)}% and rate ${Math.floor(utterance.rate * 100)}%.`)
+      })
+    } else {
+      const error = new Error(`No offline available voices installed for language ${lang}.`)
+      LOGGER.warn(error.message)
+      return Promise.reject(error)
+    }
+  }
+}
+
+/** Comparator voices based on the names.
+  * @private
+  * @param {external:SpeechSynthesisVoice} first The first voice to compare.
+  * @param {external:SpeechSynthesisVoice} second The second voice to compare.
+  * @returns {number} -1 if name of the first voice is lexicographical smaller than
+  *   the second. Zero if both names are identical. 0 otherwise.
+  */
+function voiceComparator (first, second) {
+  if (first.name < second.name) {
+    return -1
+  } else if (first.name === second.name) {
+    return 0
+  } else {
+    return 1
   }
 }